backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
supabase==2.0.2
httpx==0.25.2
google-generativeai==0.3.2
sentence-transformers==2.2.2
langchain==0.0.350
langchain-community==0.0.1
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
python-magic==0.4.27
aiofiles==23.2.1 